Character Controller & Camera Movement
The character moves with the joystick.

The character turns left and right (yaw axis) with the camera.

Make the jump button and fall by gravity.

Do the squat and crouch buttons. Pressing will adjust the hitbox and movement speed accordingly.

Swiping the screen will cause the camera to turn accordingly (doesn't work if it hits a swipe spot that doesn't empty space).